ttl expressed as no. of cache accesses (not in seconds), font-size of plots increased, logs writing on file enabled, progress bar for training added, and training+testing enabled (static)

+++ config/Config.py
@@ -69,8 +69,8 @@ class Config:
     confidence_level: float
     mc_dropout_num_samples: int
     cache_size: int
-    fixed_ttl: float
+    fixed_ttl: int
     prediction_interval: int
     threshold_score: float
     time_decay: float
-    ttl_base: float
+    ttl_base: int
